// src/pages/admin/types.ts
export interface MemberAdminPayload {
  id?: string;
  authUserId?: string | null;
  accountId: string;
  name: string;
  email: string;
  note: string;
  role: 'user' | 'admin';
  userActive: boolean;
  memberStatus: 'pending' | 'active';
  reportRequired: boolean;
  isActive?: boolean;
}